reorganizes changes for code generation
* generator - includes all the functions required to generate code
* context - Stores context structures required for code generation
* result - Stores CodeGenResult and CodeGenError
* utils - all the necessary utilities for code generation
